separate switcher widgets into different .py files
