+++ spyder/plugins/switcher/widgets/switcher.py
@@ -22,6 +22,8 @@ from qtpy.QtWidgets import (QAbstractItemView, QApplication, QDialog,
 from spyder.config.base import _
 from spyder.config.manager import CONF
 from spyder.config.utils import is_ubuntu
+from spyder.plugins.switcher.widgets.proxymodel import SwitcherProxyModel
+from spyder.plugins.switcher.widgets.switcheritem import SwitcherItem, SwitcherBaseItem, SwitcherSeparatorItem
 from spyder.py3compat import TEXT_TYPES, to_text_string
 from spyder.utils.icon_manager import ima
 from spyder.widgets.helperwidgets import HTMLDelegate
@@ -80,433 +82,6 @@ class SwitcherDelegate(HTMLDelegate):
         """
         option.state |= QStyle.State_Active
         super(SwitcherDelegate, self).paint(painter, option, index)
-
-
-class SwitcherBaseItem(QStandardItem):
-    """Base List Item."""
-
-    _PADDING = 5
-    _WIDTH = 400
-    _STYLES = None
-    _TEMPLATE = None
-
-    def __init__(self, parent=None, styles=_STYLES):
-        """Create basic List Item."""
-        super(SwitcherBaseItem, self).__init__()
-
-        # Style
-        self._width = self._WIDTH
-        self._padding = self._PADDING
-        self._styles = styles if styles else {}
-        self._action_item = False
-        self._score = -1
-        self._height = self._get_height()
-
-        # Setup
-        # self._height is a float from QSizeF but
-        # QSize() expects a QSize or (int, int) as parameters
-        self.setSizeHint(QSize(0, int(self._height)))
-
-    def _render_text(self):
-        """Render the html template for this item."""
-        raise NotImplementedError
-
-    def _set_rendered_text(self):
-        """Set the rendered html template as text of this item."""
-        self.setText(self._render_text())
-
-    def _set_styles(self):
-        """Set the styles for this item."""
-        raise NotImplementedError
-
-    def _get_height(self):
-        """
-        Return the expected height of this item's text, including
-        the text margins.
-        """
-        raise NotImplementedError
-
-    # --- API
-    def set_width(self, value):
-        """Set the content width."""
-        self._width = value - (self._padding * 3)
-        self._set_rendered_text()
-
-    def get_width(self):
-        """Return the content width."""
-        return self._width
-
-    def get_height(self):
-        """Return the content height."""
-        return self._height
-
-    def get_score(self):
-        """Return the fuzzy matchig score."""
-        return self._score
-
-    def set_score(self, value):
-        """Set the search text fuzzy match score."""
-        self._score = value
-        self._set_rendered_text()
-
-    def is_action_item(self):
-        """Return whether the item is of action type."""
-        return bool(self._action_item)
-
-    # --- Qt overrides
-    def refresh(self):
-        """Override Qt."""
-        super(SwitcherBaseItem, self).refresh()
-        self._set_styles()
-        self._set_rendered_text()
-
-
-class SwitcherSeparatorItem(SwitcherBaseItem):
-    """
-    Separator Item represented as <hr>.
-
-    Based on HTML delegate.
-
-    See: https://doc.qt.io/qt-5/richtext-html-subset.html
-    """
-
-    _SEPARATOR = '_'
-    _STYLE_ATTRIBUTES = ['color', 'font_size']
-    _STYLES = {
-        'color': QApplication.palette().text().color().name(),
-        'font_size': 10,
-    }
-    _TEMPLATE = \
-        u'''<table cellpadding="{padding}" cellspacing="0" width="{width}"
-                  height="{height}" border="0">
-  <tr><td valign="top" align="center"><hr></td></tr>
-</table>'''
-
-    def __init__(self, parent=None, styles=_STYLES):
-        """Separator Item represented as <hr>."""
-        super(SwitcherSeparatorItem, self).__init__(parent=parent,
-                                                    styles=styles)
-        self.setFlags(Qt.NoItemFlags)
-        self._set_rendered_text()
-
-    # --- Helpers
-    def _set_styles(self):
-        """Set the styles for this item."""
-        for attr in self._STYLE_ATTRIBUTES:
-            if attr not in self._styles:
-                self._styles[attr] = self._STYLES[attr]
-
-        rich_font = self._styles['font_size']
-
-        if sys.platform == 'darwin':
-            font_size = rich_font
-        elif os.name == 'nt':
-            font_size = rich_font
-        elif is_ubuntu():
-            font_size = rich_font - 2
-        else:
-            font_size = rich_font - 2
-
-        self._styles['font_size'] = font_size
-
-    def _render_text(self):
-        """Render the html template for this item."""
-        padding = self._padding
-        width = self._width
-        height = self.get_height()
-        text = self._TEMPLATE.format(width=width, height=height,
-                                     padding=padding, **self._styles)
-        return text
-
-    def _get_height(self):
-        """
-        Return the expected height of this item's text, including
-        the text margins.
-        """
-        doc = QTextDocument()
-        doc.setHtml('<hr>')
-        doc.setDocumentMargin(self._PADDING)
-        return doc.size().height()
-
-
-class SwitcherItem(SwitcherBaseItem):
-    """
-    Switcher item with title, description, shortcut and section.
-
-    SwitcherItem: [title description    <shortcut> section]
-
-    Based on HTML delegate.
-    See: https://doc.qt.io/qt-5/richtext-html-subset.html
-    """
-
-    _FONT_SIZE = 10
-    _STYLE_ATTRIBUTES = ['title_color', 'description_color', 'section_color',
-                         'shortcut_color', 'title_font_size',
-                         'description_font_size', 'section_font_size',
-                         'shortcut_font_size']
-    _STYLES = {
-        'title_color': QApplication.palette().text().color().name(),
-        'description_color': 'rgb(153, 153, 153)',
-        'section_color': 'rgb(70, 179, 239)',
-        'shortcut_color': 'rgb(153, 153, 153)',
-        'title_font_size': _FONT_SIZE,
-        'description_font_size': _FONT_SIZE,
-        'section_font_size': _FONT_SIZE,
-        'shortcut_font_size': _FONT_SIZE,
-    }
-    _TEMPLATE = u'''
-<table width="{width}" max_width="{width}" height="{height}"
-                          cellpadding="{padding}">
-  <tr>
-    <td valign="middle">
-      <span style="color:{title_color};font-size:{title_font_size}pt">
-        {title}
-      </span>&nbsp;
-      <small
-       style="color:{description_color};font-size:{description_font_size}pt">
-        <span>{description}</span>
-      </small>
-    </td>
-    <td valign="middle" align="right" float="right">
-      <span style="color:{shortcut_color};font-size:{shortcut_font_size}pt">
-         <small><code><i>{shortcut}</i></code></small>
-      </span>&nbsp;
-      <span style="color:{section_color};font-size:{section_font_size}pt">
-         <small>{section}</small>
-      </span>
-    </td>
-  </tr>
-</table>'''
-
-    def __init__(self, parent=None, icon=None, title=None, description=None,
-                 shortcut=None, section=None, data=None, tool_tip=None,
-                 action_item=False, styles=_STYLES):
-        """Switcher item with title, description, shortcut and section."""
-        super(SwitcherItem, self).__init__(parent=parent, styles=styles)
-
-        self._title = title if title else ''
-        self._rich_title = ''
-        self._shortcut = shortcut if shortcut else ''
-        self._description = description if description else ''
-        self._section = section if section else ''
-        self._icon = icon
-        self._data = data
-        self._score = -1
-        self._action_item = action_item
-
-        self._section_visible = True
-
-        # Setup
-        self.setFlags(Qt.ItemIsEnabled | Qt.ItemIsSelectable)
-
-        if icon:
-            self.setIcon(icon)
-            # TODO: Change fixed icon size value
-            self._icon_width = 20
-        else:
-            self._icon_width = 0
-
-        self._set_styles()
-        self._set_rendered_text()
-
-    # --- Helpers
-    def _render_text(self, title=None, description=None, section=None):
-        """Render the html template for this item."""
-        if self._rich_title:
-            title = self._rich_title
-        else:
-            title = title if title else self._title
-
-        # TODO: Based on width this should elide/shorten
-        description = description if description else self._description
-
-        if self._section_visible:
-            section = section if section else self._section
-        else:
-            section = ''
-
-        padding = self._PADDING
-        width = int(self._width - self._icon_width)
-        height = int(self.get_height())
-        self.setSizeHint(QSize(width, height))
-
-        shortcut = '&lt;' + self._shortcut + '&gt;' if self._shortcut else ''
-
-        title = to_text_string(title, encoding='utf-8')
-        section = to_text_string(section, encoding='utf-8')
-        description = to_text_string(description, encoding='utf-8')
-        shortcut = to_text_string(shortcut, encoding='utf-8')
-
-        text = self._TEMPLATE.format(width=width, height=height, title=title,
-                                     section=section, description=description,
-                                     padding=padding, shortcut=shortcut,
-                                     **self._styles)
-        return text
-
-    def _set_styles(self):
-        """Set the styles for this item."""
-        for attr in self._STYLE_ATTRIBUTES:
-            if attr not in self._styles:
-                self._styles[attr] = self._STYLES[attr]
-
-        rich_font = self._styles['title_font_size']
-
-        if sys.platform == 'darwin':
-            title_font_size = rich_font
-            description_font_size = title_font_size + 2
-        elif os.name == 'nt':
-            title_font_size = rich_font
-            description_font_size = title_font_size + 1
-        elif is_ubuntu():
-            title_font_size = rich_font - 2
-            description_font_size = title_font_size + 1
-        else:
-            title_font_size = rich_font - 2
-            description_font_size = title_font_size + 1
-
-        self._styles['description_font_size'] = description_font_size
-        self._styles['section_font_size'] = description_font_size
-
-    def _get_height(self):
-        """
-        Return the expected height of this item's text, including
-        the text margins.
-        """
-        doc = QTextDocument()
-        try:
-            doc.setHtml('<span style="font-size:{}pt">Title</span>'
-                        .format(self._styles['title_font_size']))
-        except KeyError:
-            doc.setHtml('<span>Title</span>')
-        doc.setDocumentMargin(self._PADDING)
-        return doc.size().height()
-
-    # --- API
-    def set_icon(self, icon):
-        """Set the QIcon for the list item."""
-        self._icon = icon
-        self.setIcon(icon)
-
-    def get_icon(self):
-        """Return the QIcon for the list item."""
-        return self._icon
-
-    def set_title(self, value):
-        """Set the main text (title) of the item."""
-        self._title = value
-        self._set_rendered_text()
-
-    def get_title(self):
-        """Return the the main text (title) of the item."""
-        return self._title
-
-    def set_rich_title(self, value):
-        """Set the rich title version (filter highlight) of the item."""
-        self._rich_title = value
-        self._set_rendered_text()
-
-    def get_rich_title(self):
-        """Return the rich title version (filter highlight) of the item."""
-        return self._rich_title
-
-    def set_description(self, value):
-        """Set the item description text."""
-        self._description = value
-        self._set_rendered_text()
-
-    def get_description(self):
-        """Return the item description text."""
-        return self._description
-
-    def set_shortcut(self, value):
-        """Set the shortcut for the item action."""
-        self._shortcut = value
-        self._set_rendered_text()
-
-    def get_shortcut(self, value):
-        """Return the shortcut for the item action."""
-        return self._shortcut
-
-    def set_tooltip(self, value):
-        """Set the tooltip text for the item."""
-        super(SwitcherItem, self).setTooltip(value)
-
-    def set_data(self, value):
-        """Set the additional data associated to the item."""
-        self._data = value
-
-    def get_data(self):
-        """Return the additional data associated to the item."""
-        return self._data
-
-    def set_section(self, value):
-        """Set the item section name."""
-        self._section = value
-        self._set_rendered_text()
-
-    def get_section(self):
-        """Return the item section name."""
-        return self._section
-
-    def set_section_visible(self, value):
-        """Set visibility of the item section."""
-        self._section_visible = value
-        self._set_rendered_text()
-
-    def set_action_item(self, value):
-        """Enable/disable the action type for the item."""
-        self._action_item = value
-        self._set_rendered_text()
-
-
-class SwitcherProxyModel(QSortFilterProxyModel):
-    """A proxy model to perform sorting on the scored items."""
-
-    def __init__(self, parent=None):
-        """Proxy model to perform sorting on the scored items."""
-        super(SwitcherProxyModel, self).__init__(parent)
-        self.setFilterCaseSensitivity(Qt.CaseInsensitive)
-        self.setSortCaseSensitivity(Qt.CaseInsensitive)
-        self.setDynamicSortFilter(True)
-        self.__filter_by_score = False
-
-    def set_filter_by_score(self, value):
-        """
-        Set whether the items should be filtered by their score result.
-
-        Parameters
-        ----------
-        value : bool
-           Indicates whether the items should be filtered by their
-           score result.
-        """
-        self.__filter_by_score = value
-        self.invalidateFilter()
-
-    def filterAcceptsRow(self, source_row, source_parent):
-        """Override Qt method to filter items by their score result."""
-        item = self.sourceModel().item(source_row)
-        if self.__filter_by_score is False or item.is_action_item():
-            return True
-        else:
-            return not item.get_score() == -1
-
-    def sortBy(self, attr):
-        """Override Qt method."""
-        self.__sort_by = attr
-        self.invalidate()
-        self.sort(0, Qt.AscendingOrder)
-
-    def lessThan(self, left, right):
-        """Override Qt method."""
-        left_item = self.sourceModel().itemFromIndex(left)
-        right_item = self.sourceModel().itemFromIndex(right)
-
-        # Check for attribute, otherwise, check for data
-        if hasattr(left_item, self.__sort_by):
-            left_data = getattr(left_item, self.__sort_by)
-            right_data = getattr(right_item, self.__sort_by)
-            return left_data < right_data
 
 
 class Switcher(QDialog):
